Improve parsing (#36)
* Clarify flymake-eslint--message-regexp

Both `any' and `not-newline' are converted to the "." regular
expression. I find the later more explicit and this use of `any' isn't
mentioned in the docstring of `rx' (in Emacs 29.2 at least).

* Fix flymake-eslint--message-regexp for multi-line messages

In some cases, eslint can produce a multi-line error message. This
happens for example for the import/no-deprecated rule [1] which shows
a user-constructed message in the error. Replacing `any' (which
matches anything but a new line) with `anychar' allows for matching
new lines within the message.

[1] https://github.com/import-js/eslint-plugin-import/blob/main/docs/rules/no-deprecated.md

* Improve JSON parsing in case of an eslint's crash

Even if "--format json" is passed to eslint, the output will be
non-json in case of a crash, e.g.,

  Error: Failed to load plugin 'foo' declared in '.eslintrc.cjs':
  Cannot find module 'eslint-plugin-foo' …

This commit makes sure that such eslint's crash doesn't prevent
flymake-eslint from reporting the error.