Use directory containing eslint config file or package.json as root
(project-root (project-current)) only works if the user has properly
configured `project` (to consider "package.json" as a
`project-vc-extra-root-markers`) or if the project has a simple
configuration (i.e., the .git/ and the package.json are in the same
directory).

This patch searches for the first directory containing a file in
`flymake-eslint-project-markers` (by default package.json and eslint
flat config files).

The (project-root (project-current)) is kept as a fallback for
non-flat eslint config files (i.e., .eslintrc.* files) and for
JavaScript files not part of an NPM package.